Add information, on when a canteen is open, based on API data
