+++ css/styles.css
@@ -16,7 +16,7 @@ body {
     min-height: 70vh;
 }
 
-.map-modal {
+.map-modal .modal-content {
     min-width: 80vw;
 }
